requirements.txt
aiobotocore==0.9.3
aiofiles==0.3.2
aiohttp==3.3.2
appdirs==1.4.3
async-generator==1.10
async-timeout==3.0.0
attrs==19.3.0
beautifulsoup4==4.8.1
botocore==1.10.58
certifi==2019.9.11
chardet==3.0.4
docutils==0.15.2
entrypoints==0.3
flake8==3.5.0
fuckcaptcha==0.1.0
idna==2.8
jmespath==0.9.4
mccabe==0.6.1
multidict==4.6.1
numpy==1.17.4
opencv-python==4.1.2.30
peewee==3.11.2
Pillow==6.2.1
pocketsphinx==0.1.15
psutil==5.4.6
pycodestyle==2.3.1
pydub==0.22.1
pyee==6.0.0
pyflakes==1.6.0
pyppeteer==0.0.25
pyppeteer-stealth==1.0.0
python-dateutil==2.8.1
PyYAML==5.1.2
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
tqdm==4.39.0
urllib3==1.24.3
user-agent==0.1.9
websockets==6.0
wrapt==1.11.2
yarl==1.3.0
